+++ tests/Python26SocketServer.py
@@ -128,7 +128,6 @@ BaseServer:
 
 __version__ = "0.4"
 
-
 import socket
 import select
 import sys
@@ -329,12 +328,11 @@ class BaseServer:
         The default is to print a traceback and continue.
 
         """
-        print '-' * 40
-        print 'Exception happened during processing of request from',
-        print client_address
+        print('-' * 40)
+        print('Exception happened during processing of request from %s' % client_address)
         import traceback
         traceback.print_exc()  # XXX But this goes to stderr!
-        print '-' * 40
+        print('-' * 40)
 
 
 class TCPServer(BaseServer):